backend/requirements.txt
fastapi==0.109.2
pydantic==2.5
pydantic_settings==2.2.1
uvicorn==0.24.0.post1
SQLAlchemy==2.0.29
python-dotenv==1.0.1
asyncpg==0.29.0
aiogram==3.4.1
jwt==1.3.1
python-jose==3.3.0
httpx==0.27.0
python-multipart==0.0.9
anthropic==0.21.3
pyrogram==2.0.106
